+++ eqa/lib/state.py
@@ -2,7 +2,7 @@
 
 """
    Program:   EQ Alert
-   File Name: eqa_settings.py
+   File Name: eqa/lib/state.py
    Copyright (C) 2019 Michael Geitz
 
    This program is free software; you can redistribute it and/or modify
@@ -16,41 +16,38 @@
    You should have received a copy of the GNU General Public License along
    with this program; if not, write to the Free Software Foundation, Inc.,
    51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+   Parse and react to eqemu logs
 """
 
-import sys
-import os
-import logging
-import time
-import datetime
-import yaml
-import pwd
-import grp
-from collections import namedtuple
 
-def usage():
-  """Print some helpful things"""
-  print('ya done goofed')
-  sys.exit(2)
+class EQA_State:
+  def __init__(self, char, chars, zone, loc, direction, afk, server):
+    self.char = char
+    self.chars = chars
+    self.zone = zone
+    self.loc = loc
+    self.direction = direction
+    self.afk = afk
+    self.server = server
 
+  def set_char(self, char):
+    self.char = char
 
-def timestamp():
-  """Returns a neat little timestamp for things"""
-  unixstamp = int(time.time())
-  timestamp = datetime.datetime.fromtimestamp(int(unixstamp))\
-      .strftime('%Y-%m-%d_%H:%M:%S')
-  return str(timestamp)
+  def set_chars(self, chars):
+    self.chars = chars
 
+  def set_zone(self, zone):
+    self.zone = zone
 
-def eqa_time():
-  """Returns message timestamp HH:MM:SS.ff"""
-  return datetime.datetime.now().strftime('%H:%M:%S.%f')[:-4]
+  def set_loc(self, loc):
+    self.loc = loc
 
+  def set_direction(self, direction):
+    self.direction = direction
 
-def log(message):
-  """Effectively just for timestamping all log messages"""
-  logging.info('[' + timestamp() + ']: ' + str(message))
+  def set_afk(self, afk):
+    self.afk = afk
 
-
-if __name__ == '__main__':
-  main()
+  def set_server(self, server):
+    self.server = server
